dj_app/requirements.txt
Django==3.2.7
django-redis==5.0.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
redis==3.5.3
django-cors-headers==3.8.0
djangorestframework==3.12.4
channels==3.0.4
channels-redis==3.3.0
pytest-django==4.4.0
slack-sdk==3.10.1
azure-servicebus==7.3.3